Update - in AnalysisMethods, Operator, MatrixOperator: Summenberechnung korrigiert (Produkte fehlen noch!).

+++ src/abstractexpressions/expression/classes/Operator.java
@@ -1850,7 +1850,7 @@ public class Operator extends Expression {
                     BigInteger.valueOf(ComputationBounds.BOUND_OPERATOR_MAX_NUMBER_OF_MEMBERS_IN_SUM_OR_PRODUCT)) <= 0) {
 
                 // Dann kann man die Summe explizit ausschreiben.
-                return AnalysisMethods.sum(summand, (String) this.params[1], lowerLimit.intValue(), upperLimit.intValue());
+                return AnalysisMethods.sum(summand, (String) this.params[1], lowerLimit, upperLimit);
 
             }
         }
